Address review: use pandas sorting, rank-based reverse, drop float cast
- Slow path: use alifestd_mark_node_depth_asexual + pandas sort_values
  instead of reimplementing depth calc and using sorted() with lambda
- Fast path: use rank-based reversal (-argsort(argsort())) instead of
  negating criterion values, safe for any sortable type
- Remove .astype(float64) cast since rank-based approach handles any type

+++ phyloframe/legacy/_alifestd_sort_children_polars.py
@@ -106,12 +106,7 @@ def alifestd_sort_children_polars(
         .to_numpy()
     )
     criterion_values = (
-        phylogeny_df.lazy()
-        .select(criterion)
-        .collect()
-        .to_series()
-        .to_numpy()
-        .astype(float)
+        phylogeny_df.lazy().select(criterion).collect().to_series().to_numpy()
     )
 
     logging.info(
